📁: 파일, 폴더 삭제 및 경로 수정
