--- src/components/SocialLoginBtn/social-loginbtn.scss
@@ -1,48 +1,0 @@
-@mixin link-module ($url, $color) {
-  font-family: 'Spoqa Han Sans Neo';
-  background-image: url($url);
-  background-repeat: no-repeat;
-  background-position: 17px center;
-  border: 1px solid $color;
-  border-radius: 44px;
-}
-
-.list-social-login {
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-  
-  li {
-    width: 322px;
-    margin-bottom: 10px;
-    &:last-child {
-      margin-bottom: 0px;
-    }
-
-    a {
-      display: block;
-      padding: 13px 80px;
-      text-align: center;
-      text-decoration: none;
-      line-height: 18px;
-      font-weight: 400;
-      font-size: 14px;
-      background: #FFFFFF;
-      color: #767676;
-      box-sizing: border-box;
-      border-radius: 5px;
-    }
-
-    .link-kakao {
-      @include link-module("../../assets/images/kakao-logo.svg", #F2C94C);
-    }
-
-    .link-google {
-      @include link-module("../../assets/images/google-logo.svg", #767676);
-    }
-
-    .link-facebook {
-      @include link-module("../../assets/images/facebook-logo.svg", #2D9CDB);
-    }
-  }
-}
