+++ src/pages/Login/LoginSection.scss
@@ -1,9 +1,11 @@
-@import '../../styles/color.scss';
+@import '../../Components/scss/default.scss';
 
-.content-login-section {
+.section-login {
   position: absolute;
   bottom: 0;
-  width: 100%;
+  left: 0;
+  right: 0;
+  margin: 0 auto;
   padding: 50px 0 100px;
   border-radius: 20px 20px 0 0;
   background-color: $white;
