+++ src/pages/Login/SocialLoginBtn.jsx
@@ -1,0 +1,16 @@
+import FacebookBtn from './FacebookBtn';
+import GoogleBtn from './GoogleBtn';
+import KakaoBtn from './KakaoBtn';
+import './SocialLoginBtn.scss';
+
+function SocialLoginBtn () {
+  return (
+    <ul className="list-socialbtn">
+      <KakaoBtn/>
+      <GoogleBtn/>
+      <FacebookBtn/>
+    </ul>
+  )
+}
+
+export default SocialLoginBtn;
