--- src/styles/color.scss
@@ -1,2 +1,0 @@
-$login-background-color: #EA7F42;
-$white: #FFFFFF;
